Add excision property to MVFv4
This provides a float32 value between 0.0 and 1.0 per visibility that
indicates the fraction of CBF visibilities that were excised by the
SDP ingest pipeline when forming the SDP L0 visibility. It is affected
by dataset selection like vis, flags and weights, and also calculated
lazily.

The excision is derived from the unscaled weights, which contain the
number of actual accumulations that make up each SDP visibility. These
counts are quantised to fit into a byte by the lossy weight compression
scheme in ingest, where 255 is mapped to the maximum count per channel.
The stored weights therefore have compression artefacts:

  1. If an SDP dump contains e.g. 4 CBF dumps (a common case),
     the weights byte might contain values of 255, 191, 127, 64 and 0.
     Dividing these by 255 do not result in exact multiples of 1/4.

  2. If an SDP dump contains more than 256 CBF dumps (e.g. in the
     sdp_l0_continuum stream, which is a heavily frequency-averaged
     version of sdp_l0), some precision will be lost.

The first artefact can be fixed by forcing the unscaled weights to be
integer multiples of the accumulations per CBF dumps, since each SDP L0
dump can only contain an integer number of CBF dumps. The second
artefact cannot be fixed, but it is also less common.

If the data is good quality, the excision can be expected to be 0.0,
although it can also happen due to rounding error in case 2 if below
1 / 512. The excision can also be 1.0 if the entire visibility was lost
and replaced by a zero value, or due to case 2 rounding errors if it
is larger than or equal to 511 / 512.

This addresses JIRA ticket SR-1979.