Split out separate utility classes for Minio
This will make it practical for other packages to reuse the code.

It is deliberately designed not to use the SkipTest exception, which is
specific to nosetests, so that it can easily be integrated into other
test frameworks (pytest).

The version test has been removed and left to documentation: newer
versions of minio have a different way of querying the version, and it
didn't seem worth trying to support both flavours.