Add basic Van Vleck routines to correct autocorrs
Add basic routines to perform quantisation (Van Vleck) correction of
visibilities. This only focuses on autocorrelations (ie power/variance)
and not crosscorrelations, which is appropriate for correlators with
many quantisation levels like the MeerKAT 8-bit correlator.

This only corrects the mean autocorrelation. The uncertainty of the
autocorrelation is a function of its mean and is therefore also
improved, although a proper characterisation of the uncertainty
requires a Bayesian treatment which is not done here. This is also
true of the uncertainty of the crosscorrelations.

Under the assumption of a proper complex Gaussian signal, I calculate
the variance of the quantised voltage signal (ie what the correlator
estimates) for a given variance of the unquantised signal and then
construct a lookup table that will allow us to go in the opposite
direction (quantised -> unquantised). The lookup table is suitable for
linear interpolation.

This is based on research dating back to 2012. I ditched the spline
interpolation that caused some ringing at the edges of the domain and
stripped out the Bayesian machinery to get going. Once the splines were
gone, I could also get rid of a dependency on SciPy by replacing
scipy.stats.norm.cdf with a Numba vectorised version of math.erf.
Interestingly enough the new implementation is also twice as fast.