Hook up Van Vleck to ChunkStoreVisFlagsWeights
The basic Van Vleck correction can now be turned on and off by passing
the `van_vleck` parameter to `ChunkStoreVisFlagsWeights` (default=off).
This adds a layer of linear interpolation to the visibility dask graph.
The quantisation levels are currently hardwired to the MeerKAT 8-bit
requantiser output.

This addresses JIRA ticket SPR1-674.