Print error response content for better debugging
An S3 server response with a status code in the 400-600 range is
typically turned into an exception. The exception message contained
the HTTP status code and reason as well as the request method and URL,
but not the actual response content. This content is frequently useful
for debugging, for example by mentioning that a 404 is due to a missing
bucket and not the object itself.

The response content is only added to the exception message if it is
an XML or plaintext document, making it sensible to print out. Also
pretty-print the XML to make the error barf more palatable.

In the process, discover that `KeyError` always repr's its argument,
which defeats the pretty XML. Quoting Python's Objects/exceptions.c:

/* If args is a tuple of exactly one item, apply repr to args[0].
    This is done so that e.g. the exception raised by {}[''] prints
        KeyError: ''
    rather than the confusing
        KeyError
    alone.  The downside is that if KeyError is raised with an explanatory
    string, that string will be displayed in quotes.  Too bad.
*/

Change the string representation of `ChunkNotFound` back to a plain
str() because we use explanatory strings all the time... Maybe I should
invest in a WordyKeyError :-)

While we are at it, prefix the token proxy logs to keep track of things.