Add support for S3ChunkStore URLs with paths
It will be useful for the unit tests to have a chunk store that is
relative to a single existing bucket. This is achieved by adding a path
to the URL that initialises `S3ChunkStore`. Most of the code is already
fine with this as it uses fully specified URLs (store URL + chunk name).

The one exception is bucket creation. Previously we assumed that the
first component of the array name is the bucket name and the store URL
is an endpoint / network location, but now the bucket name may reside
in the store URL instead. Extract the bucket URL from the full URL
instead and adapt the internal `_chunk_url` and `_create_bucket`
methods. The `mark_complete` method is also standardised to use the
top-level `create_array` method instead.