Rework S3 unit tests to check RDB support
The goal is to unit test `TelstateDataSource.from_url` for RDB files
stored in S3. Since the S3 tests heavily depend on helper classes to
drive MinIO and to check authorisation, it seems best to include the
data source tests with the chunk store tests.

The `TelstateDataSource.from_url` method constructs its own S3ChunkStore
to load the RDB file, based on the RDB URL and kwargs, instead of
relying on the `chunk_store` parameter. This is useful to mix and match
different stores for the metadata and data, but requires a rework of
`TestS3ChunkStore.from_url` to expose the url and kwargs too. A side
benefit is that tests can now override timeouts, retries and tokens.

Include the single bucket in the unit tests in the store URL, so that
everything is relative to it. This simplifies authorisation, which can
use a single token for everything. There is no more need for a custom
`TestS3ChunkStore.array_name`. In fact, we might be able to remove it
entirely from `ChunkStoreTestBase` since the S3 version was the only
non-trivial use case.

The test coverage of datasources.py now stands at 91% - good enough
for now.