Simplify S3ChunkStore creation in unit tests
Remove `from_url()` as it does not offer much. Instead, promote
the fittingly named `prepare_store_args()` to the public API and
simplify it in the process. There is no need for a separate
`authenticate` parameter, since we could just set `credentials`
directly (thanks to `setdefault`) in the case of `TestS3ChunkStore`,
while the token version of the class is never used without
authentication. Let the token version call the non-token super()
to share some code.