Rework bucket organisation and authorisation
Instead of forcing everything into one bucket, use a two-pronged
approach:

- Standard tests end up in a single standard bucket to avoid invalid
  bucket names like `x` and `big_y` (forgot about that reason!).
- Advanced tests like `test_rdb_support` can use a standard prefix
  styled as a CBID, allowing realistic buckets like `1234567890` and
  `1234567890-sdp-l0`.

This simplifies token authorisation, since the BUCKET and PREFIX are
known up front. The store URL returns to being an endpoint (but support
for extra paths remain), so that paths are absolute.

The suggestion strings are now also tacked onto the end of the array
name, since there will always be a final path component / filename to
book-end it (the regex looks for /suggestion/). Inserting it between
the bucket and array name is problematic since the array name may be
equal to the bucket name (think of the typical RDB files like
`/1234567890/1234567890_sdp_l0.rdb`).