Increase test time and improve log message
The test ran out of time, so ease the time limit a bit.

The token proxy log message did not distinguish between a fresh
suggestion, with initial_request_time == now, and a stale or absent
suggestion, with no initial_request_time that also defaulted to now.
Rather default that to now + 1, which will result in a negative time
offset in the logs, thereby distinguishing these two cases.