Rename prepare -> _put_chunk
It's really a private helper method. There is already a
`prepare_store_args` which can cause confusion. And it's a bit weird
to see a `get_chunk` without a `put_chunk`.

Extend and standardise the timeouts of persistent error tests, since
it seems like failure takes a bit longer, in general.