Introduce a PlaceholderChunk class
get_chunk_or_none is replaced by get_chunk_or_placeholder. The advantage
of the placeholder is that it can remember the shape and dtype that were
expected, removing the need to embed them into the graph. It also
implements `__getitem__`, in theory allowing a Dask array that contains
placeholder chunks to be sliced.