Switch to PyJWT 2 and Python 3.6
The latest PyJWT has some breaking API changes. Two are relevant:

- The `jwt.encode` function now returns `str` and not `bytes`.
- The `jwt.decode` function won't verify any claims without verifying
  the signature first.

The second one is a bit of a pity. In PyJWT 1 there was a `verify`
parameter to `jwt.decode` that defaulted to True and enabled claim
verification. This was separate from the `verify_signature` option
(only serving as its default), so we could verify claims without
verifying the signature (which would be impossible in katdal without
the server secrets).

In PyJWT 2 there is no `verify` parameter and `verify_signature` serves
both purposes, so we are forced to do our own expiration time check.

Since PyJWT 2 requires Python 3.6+, the same now goes for katdal.