Run pyupgrade --py36-plus
Upgrade code automatically to Python 3.6 syntax. By far the most common
changes are:

- Old-style formatted strings to f-strings
- Remove inheritance from `object`
- Use `super()`
- Replace `set([...])` with `{...}`
- Remove 'u' literal string prefixes

I've edited the output to align whitespace and to preserve the %
formatting in logger output (in those crufty old places where I first
assembled the formatted string instead of passing the arguments to the
logger as one should). In a handful of places I also simplified the
format strings further.