Manual formatted string conversion
Convert as many formatted strings as possible to f-strings. When it is
ugly or inconvenient, at least convert `%` to `.format()`. Also let
logger outputs be formatted inside the logging system by turning the
parts into arguments (missed a few of these from the old days).

The main part left untouched is DataSet.__str__, for which I have
special plans.