Add index parameter to katdal.open for visdatav4.
It can be used to slice out a contiguous subrange of time or frequency.
Apart from slicing the data, it also slices timestamps and the
SpectralWindow. There may be other related metadata that I've missed.

The slices are restricted to contiguous forwards slices (step=1) to
avoid extra complications with adjusting dump_period / channel_width.