KAT-1196 Allow non-radec targets (planets) in MS
Very early on (see 73546ac) I decided to disable non-radec targets
because I did not understand the MS format well enough to be sure that
it would work. The FIELD subtable stores directions that may vary as
polynomials over time and it all seemed very complicated.

Fast-forward 10 years to the Great Conjunction and a subsequent clamour
for MSes with Jupiter... Having studied the `fixplanets` CASA task and
the `planets_6cm` demo dataset, I came to the conclusion that it is
much ado about nothing. The standard procedure in CASA picks the first
timestamp associated with a field and determines the (ra, dec)
from the (az, el) POINTING table in the MS. In other words, just like
`target.radec(first_timestamp)` :-)

Looking at `fixplanets`, there is also an option to use the Horizons
ephemeris as an alternative. If we figure this out someday it might be
more accurate for Solar System objects, but for now this trick works
for all non-radec targets. :tada: