KAT-1196 Allow scans in MS
While we are at it, also allow scans (and not only tracks) into an MS.
The FIELD and SOURCE tables will be populated from the target (ra, dec)
and the only difference is that the antenna pointing won't match that.
But we are not storing the POINTING table yet, so it should not make
a difference. This at least provides users with the ability to look at
scan data and figure it out themselves :-)