SPR1-897 Allow nested DaskLazyIndexers
The main use case is to slap some more transforms onto an existing
indexer (think `flags` piggy-backing on `raw_flags`) but it also allows
N-stage indexing. The first indexer's `dataset` Dask array becomes the
original dataset for the second indexer (lazily), and an additional
selection and transform set will then be applied to it to form the
`dataset` of the second indexer.