SPR1-897 Remove DaskLazyIndexer.add_transform
It makes DaskLazyIndexer mutable, which is a problem once you start
nesting indexers. If you invoke `add_transform` on an upstream indexer
after you have already materialised a downstream indexer, then the
downstream indexer won't get those changes.

The `add_transform` method is not used in katdal at all... It was added
in #180 in anticipation of the impending applycal, but applycal ended up
using `da.map_blocks` and friends instead.