SPR1-1138 New applied_delay virtual sensor
Extract the CBF antenna_channelised_voltage delay sensor, which
represents the delay tracking updates applied in the F-engine, to
a proper (virtual) katdal sensor. The true timestamps representing
the start of each delay update period are ADC sample counts contained
in the values themselves, not the traditional sensor timestamps. These
typically differ by the command lead time of around two seconds.

Use the reported delay rate to interpolate the delay within each update
period, for extra accuracy. Do this by inserting an extra sample
computed with this delay rate just before the start of the next update
period. The normal sensor interpolation machinery will then do linear
interpolation across the bulk of each update period.

The one niggle is the need to access CBF attributes to find the
underlying delay sensor and to convert the timestamps. These attributes
are turned into single-valued sensors at the moment, which is a bit of
a kludge but allows the sensor function to remain outside the dataset
class. An alternative is to pass the attributes to the sensor function
via functools.partial at registration time, but that is also not ideal.