SPR1-1138 New applied_phase virtual sensor
Extract the piecewise linear fringe phase function applied in the
F-engine as part of delay tracking to a proper katdal sensor. The
`applied_phase` and `applied_delay` sensors are constructed together
and cached, given their intertwinedness in the CBF sensor.

Use the reported fringe rate to interpolate the phase within each delay
tracking update period, for extra accuracy. The interpolated angles are
left as is and not wrapped to the range of -pi to pi radians. It is
always possible to add a transform for that.