SPR1-1138 New applied_gain virtual sensor
This reports the requantisation gains applied in the F-engine using
the ?gain command. In essence we add an alias and a transform.
The transform turns the lists in the CBF sensor into ndarrays of the
appropriate complex64 type.

It also rounds the real and imaginary parts because the F-engine casts
them to int16 before applying them, while the CBF sensor only seems
to report back the CAM request. More on this in SR-1952 and SPR1-1178.
This should be more accurate if you want to undo the gain.