SPR1-1134 Rework name and url for older datasets
Call the base part of the HDF5 filename the "capture block ID", since
it is also a Unix timestamp at the start of the observation. This then
becomes the new name for MVFv1 and MVFv2 datasets. The MVFv3 format
actually has a stream name which defaults to "sdp_l0", so use that.

The actual filename is now converted to a "file://" URL and exposed as
`dataset.url`, but first it is resolved to an absolute path. This is
because file-based URLs can't really handle local paths (try opening
it in a browser...).