SPR1-1134 Redo filename to URL conversion
The existing approach had two issues:
  - `Path.resolve()` does not mesh well with Windows, since it turns
    forward slashes into backslashes. Better to use `Path.absolute()`.
  - RDB filenames were converted to file:// URLs with relative instead
    of absolute paths.

Move the filename -> URL conversion to the new `parse_url_or_path`
helper function and centralise the conversion in `DataSet.__init__`,
which once again accepts a `url` parameter (technically url_or_path...).
Add a unit test for the helper function.

This is needed by `TelstateDataSource.from_url` as well because once
you have a file:// URL, the path is forced to be absolute. It is
therefore too late to pass a reconstituted URL to DataSet to let it
extend the path.