SPR1-766 Verify that S3 bucket is not empty
Whenever a chunk is not found, go and check the associated bucket.
If the bucket is missing or empty, raise `StoreUnavailable` instead
since this is a fatal error. In the past katdal would happily return
a bucketload of zeros with no complaints.

The bucket exists but is empty if the dataset has been archived to tape.
A completely missing bucket is unlikely, but check for it anyway.

We perform the check with every missing chunk because that is simpler
than tracking which buckets we verified. The overhead should be low,
since we typically either have all chunks missing (the aim of this fix)
or only a handful of chunks missing. Besides that, 404s are quite quick
compared to the downloading of actual chunks.

We used to list the objects in a bucket in the distant past, but it was
not that useful in the end. This marks a very lean return of that smoke
test :-)