SPR1-766 Fix failing unit tests
The `test_chunk_non_existent` test typically does `get_chunk` on an
empty chunk store. This is now an issue for `S3ChunkStore`, since an
empty bucket replaces the benign `ChunkNotFound` error with the more
severe `StoreUnavailable` error. Circumvent this by putting a tiny
crumb in the bucket before the test proper starts.

The `test_public_read` test used to raise `ChunkNotFound` for an
unauthorised chunk, because 403 is also mapped to 404 for chunks due to
Ceph RGW limitations. This then triggers a bucket verification, which
results in a raw 403 and subsequent `StoreUnavailable` error. Let the
test pick up this error instead.

The 403/404 Ceph malarky will be revisited in another PR.