SPR1-766 Cache successfully verified buckets
I initially resisted adding more state to the S3 store object. However,
Bruce pointed out the possibility that an ingest node can fail, which
would result in 1/4 of the data going missing. The large number of
bucket verification requests may then cause performance issues. Instead
of checking the performance, just avoid this scenario all together...

If a bucket is not empty, add its url to an internal set. The next time
around it won't be checked again. This assumes that chunks are not
removed from a bucket while you are accessing it with katdal, which is
a reasonable assumption for the MeerKAT archive. As long as tokens are
not issued for buckets that are about to be archived to tape, this
assumption should hold.

Add a (somewhat white-box) test to check that the bucket is added to
the internal cache if it is non-empty.