SPR1-930 Fix floating-point warning in unit tests
When interpolating bandpasses with `complex_interp`, the `left` and
`right` parameters contain NaNs (INVALID_GAIN). These are passed on to
`np.unwrap`, which triggers an "invalid" floating-point error. This
emerges as a RuntimeWarning in the tests. Rather suppress this error,
since the answer for `np.unwrap` will still be the desired NaN.