SPR1-1450 Depend on katpoint 1.0a1
Let the games begin :-) Since each successive alpha release will
probably break the API yet again, it is best to depend on a hard-wired
version for now. The latest ska-telescope/katpoint master already
contains incompatible interfaces based on Astropy units but is not
released yet, which is why we settle for 1.0a1.