SPR1-1737 New mvf_copy.py script
This makes a local copy of an MVF4 dataset (local or remote). In the
process it also allows stripping out correlation products. It currently
only supports extracting the autocorrelations via the `--corrprods=auto`
script option. The filtered RDB file and chunks are stored in a new
subdirectory (by default "./archive").

The script has much in common with mvf_rechunk.py and might eventually
merge with it.