SPR1-1737 Don't modify original telstate
It is unsafe to modify telstate in place, since it may be the live
Redis database. Rather copy the original telstate as is and put the
overridden keys in a separate memory-backed telstate. The RDB loader
and memory backend are OK with duplicate keys in the RDB file, so the
simplest solution is just to write the extra overrides at the end.