SPR1-1312 Redo last dump when continuing MS
An MS dump is not written atomically by `mvftoms`. Each dump typically
adds as many rows as there are baselines to the main MS table.
The table is first padded out with zeroes via `table.addrows()`, and
then the columns in this new section are filled in one at a time via
`table.putcol()`. (We could consider `tablerow.put` as an alternative,
but this does not solve the non-atomic dump issue - it just has missing
rows instead of missing columns.)

If the script crashes, we therefore frequently end up with an
incomplete final dump in the MS, where some columns are complete and
others are zeroed. If the MS exists, rather redo this dump to be sure.
This has the added benefit that you will always write something to disk,
even if the MS was already complete, which avoids crashing the script
due to writing no visibilities.

This requires the `addrows` call in `ms_extra` to be modified, since we
don't want to add new rows for the last existing dump.