SPR1-1493 Defer turning delays into phase slopes
We want to apply corrections to delay tracking, which involves a new
delay per input at each dump. Expanding these delay corrections to
phase slopes in the correction sensor can be prohibitively expensive
(e.g. 4000 dumps x 128 inputs x 32k channels = 134 GB!).

Change the delay correction sensor to store the corrections in the form
of delays instead of complex gains. For example, 4000 dumps x 128 inputs
will only need 4 MB of storage. These delays are basically the same
as those produced by the cal pipeline, just reindexed and with NaNs
removed. That's because the cal pipeline works with raw phase slopes,
so the delay solutions are already in the form of corrections.

Postpone the conversion to complex gains to the "channel map" function,
which is now rebranded as a function that maps a cal product onto the
frequency axis of the visibility data. This gets called inside
`calc_correction_per_corrprod` as part of the applycal map_blocks
function - in other words, now right before the correction is applied
to the data.