SPR1-1177 Refactor utility functions to own module
This is in preparation for the `align_scans` function which will also
be shared among the various data formats. Instead of dumping it into
`dataset.py`, move the lot into its own utils (*shudder*) module.

This commit purely moves the code across to facilitate git tracking,
but it is not hooked up yet.