SPR1-1177 New align_scans function based on MVF4
Move the logic that aligns the scan, label and target sensors to its
own utility function. This will allow unit tests to verify and correct
its behaviour. Use the MVF4 version, although the idea is that at least
MVF2 and MVF3 will also share this function, once we have merged the
corner cases.

Deep copy the original sensors to avoid modifying them in place. This
will allow us to extract the inputs and outputs of this function after
it was run, which simplifies getting data for unit tests.