SPR1-1944 Fix getem usage in put_dask_array
It seems that `da.map_blocks` is a better fit for `put_dask_array`
than constructing the graph ourselves, since we are effectively calling
the same `put` function on each chunk and grouping together the return
values. This also simplifies the various array names in use.

Turn `_scalar_to_chunk` into the `put_map_blocks` function, since this
is the only place where it is used. It massages the scalar output of
`put_chunk` into an ndarray of the appropriate dimensions to allow
reassembly into a dask array. In addition, add a step to extract the
required `slices` specification from the `block_info` parameter.