SPR1-1944 Check overheads for dask transfer
Add a unit test to check both dask array setup time and the time to
transfer a dask array from one in-memory chunk store to another.
This minimises the actual data transfer time and focusses instead
on the overhead of the ChunkStore / dask machinery (which we are
busy tweaking).

The thresholds should be quite generous and only meant to catch
egregious regressions.