SPR1-1944 Move put_map_blocks to module level
This is necessary to avoid massive tokenization costs on setting up
the dask array. Map_blocks insists on calling `da.core.tokenize` on
its arguments, which includes the put function. This can't be
overridden by either the `name` or `token` arguments.

Functions are tokenised by first pickling them. If the function
is local or a bound method, this step includes the entire store object
(self) as well, resulting in a serialised string containing 100s
of megabytes just to produce a hash.

This is avoided by moving the `put_map_blocks` function to the module
level, which makes it more picklable. The store object is passed as
a keyword parameter, which allows it to be tokenized explicitly instead
of implicitly. Instead of precalculating e.g. the slice offsets, this
now has to happen on each call, unfortunately.