SPR1-1944 Implement put_dask_array with blockwise
This is an attempt to improve on the slight performance drop of
`da.map_blocks`. Taking our cue from the internals of the new
`da.graph_from_arraylike` function, we call `dask.blockwise.blockwise`
directly. This allows us to construct a custom graph with three
arguments to our putter function (adding the actual chunk to the usual
two getter arguments). We need this because the returned graph is
now a high-level one that is harder to modify than an old-school dict.

This improves performance to equal that of the original getem version,
even surpassing it when there are many chunks. The only downside is
that classes like `ArraySliceDep` are brand new and might still change
yet again in the near future, having been added in the recent
dask/dask#7417 PR.