Revert "SPR1-1944 Implement put_dask_array with blockwise"
Go back to `da.map_blocks` implementation of `put_dask_array` since
this works on all versions of dask, at the cost of a small penalty
in performance. The main visibility writer calls `put_chunk` directly
anyway, so `put_dask_array` is hardly used and its performance is
therefore not paramount. We keep the `blockwise` commit for posterity.

This reverts commit 5308783bd024fee0612de41bca6cb34fbb106271.