SPR1-1944 Change get_dask_array to use from_array
We're back at :func:`dask.array.from_array`. This needs an array-like
object with .shape, .ndim and .dtype, and an overridden `__getitem__`
or custom getter function.

Use the object to store the constant arguments needed for `get_chunk`
(`array_name`, `dtype` and extra `kwargs`), so dask does not have to
pass it along anymore. We no longer need the initial `functools.partial`
to insert the `dtype`, so this extra layer comes at little cost. It
also means that the `inline_array` option to `from_array` is no longer
required (it works either way).

The performance pretty much matches what we had with `getem`, so all's
good (~1800 MBps for 100 chunks and ~1540 MBps for 1000 chunks, 800 MB
in-memory arrays and DictChunkStore). It would have been nice to
include `__setitem__` and use this object for `put_dask_array` as well
(in tandem with `da.store`),  but there is no way to return error info.