Refactor part of get_dask_array into _prune_chunks
This code prunes the chunk specification to throw out any chunks that
are outside the slices in the index, and then adjusts the index to
select the same range of data as before. It is a bit hard to follow
and therefore deserves its own unit test. It also clutters up the
`get_dask_array` function, so it is a perfect candidate for a helper
function.

Having a unit test revealed that the code is actually broken for
non-zero offsets :-) This is not surprising, since we have probably
never used the combination of non-zero offsets and a non-empty index.
Offsets were meant for put_dask_array
(see https://github.com/ska-sa/katdal/pull/112#discussion_r162270558)
in the context of vis_writer but then never used... They were added to
get_dask_array only for unit tests. Add an assertion to catch any of
these cases.