SPR1-1944 Fix get_dask_array quirks
The `_ArrayLikeGetter` scheme only works properly on dask 2022.03.0,
and then only because of a regression (see dask/dask#8753).

When using the `index` parameter of `get_dask_array` to slice the array
into a smaller one, the dask `fuse` graph optimisation step merges
slices of slices into single slices, but this makes it incompatible with
the chunk store, which expects very specific slices. That is, the
chunk store dask array has a fixed set of chunks and is not arbitrarily
sliceable as expected by `fuse`.

The solution is to specify a custom getter to `da.from_array`, even
though it is the expected `_ArrayLikeGetter.__getitem__`. This prevents
further optimisation of those getter tasks.

However, the plot thickens... The custom getter is not called by
the latest dask 2022.04.0 unless you also set `inline_array=True`.
This parameter first appeared in dask 2021.01.0, which also happens
to be one of the last releases that still supports Python 3.6, so
it's all good.