SPR1-1944 Restore support for older dasks
Until we can bump the dask version in our base container to >=2021.1.0
it is best to restore support for dask >= 1.2.1.

Only call `da.from_array` with `inline_array=True` if it has the
keyword argument. Use dask's own `has_keyword` utility function for
extra marks.

Add `asarray` and `lock` parameters to our custom getter but ignore
them. This fixes things on dask 1.2.1, which calls the getter with four
parameters.

This commit has been extensively unit-tested on Python 3.6 and 3.7 with
the oldest dependencies from setup.py installed.