SPR1-1177 Quick and dirty categorical from string
Add helper functions that convert a string into a basic categorical
sensor and vice versa. Sensor values are single non-blank characters,
and events are their locations within the string. The string

  'A      B      C  A       C     '

therefore results in a sensor with values 'A', 'B' and 'C' and events
at 0, 7, 14, 17 and 25, with a total length of 31.

This will be very helpful when writing unit tests to check sensor
alignment functions, as the alignment can be quickly visualised with
strings.