SPR1-1944 Test lazy indexer on top of chunk store
Verify that `DaskLazyIndexer` works on top of the chunk store
(especially the dask bits). Include this test with `VisFlagsWeights`
so that it closely resembles the action of `VisibilityDataV4.vis`.
In particular, test the case where a single dump is selected as in
mvftoms, as this applies extra dask optimisation in `DaskLazyIndexer`.

This reproduces the error in #354.