SPR1-1944 A bad chunk ID type is a TypeError
Previously, if the `slices` parameter found on many ChunkStore methods
(aka the "chunk ID") was not a sequence of unit-stride `slice` objects
with explicit `start` and `stop` attributes, a `BadChunk` was raised.
The base class of `BadChunk` is `ValueError`, however, while an
incorrect type for `slices` should be a `TypeError` instead.

This came to a head with the recent rework of the dask machinery in
the chunk store, in quite a convoluted way... When a single dump is
selected in `DaskLazyIndexer`, an extra step of `dask.optimization.cull`
is performed. This tries to find some shortcuts during `__getitem__`
which ends up calling `__getitem__(0)` on the new `_ArrayLikeGetter`
class. Dask cull would have tried something else if this raised a
KeyError, IndexError or TypeError, but BadChunk sails right through.

Turn an invalid `slices` parameter into a plain `TypeError` instead of a
`BadChunk` to make dask happy and to clarify the purpose of `BadChunk`.

Another solution would be to base `BadChunk` on both `ValueError` and
`TypeError`, leaving `ChunkNotFound` for `KeyError`. This seems
dodgy, however, since `TypeError` is more indicative of a programming
or API bug which should be fixed, while `BadChunk` indicates some
mismatch between a requested and actual chunk as a safety check.

While we are at it, fix the definition of unit stride in `_prune_chunks`
to match the one in `chunk_metadata`. Update docstrings and unit tests.

This fixes #354.