SPR1-2367 Remove IMAGING_WEIGHT column from MS
This column is initialised full of zeroes and just takes up space
(4 bytes per main table entry, i.e. time-frequency-baseline cell, so
about 6-12% overhead). It can be recreated in CASA at a later stage.

The column is a leftover of good old blank.ms, which included it along
with MODEL_DATA and CORRECTED_DATA. The latter two columns were made
optional in the sneaky 41af4b8 as part of #75 in version 0.9, but this
one was missed.

Also remove the commented-out code that has tagged along forever :-)