Override targets from file or catalogue
Add a `target_overrides` parameter to `katdal.open`. It accepts a
katpoint Catalogue or CSV file which provides alternative target
descriptions. These descriptions override the corresponding
`katpoint.Target` objects with matching target names.

The original (user) target tags are preserved, however, since it is
possible to have the same target with different tags in the data set,
e.g. both as imaging target and calibrator. This avoids the two
versions collapsing to a single target.

This feature is mainly used to change the target coordinates and/or
its flux density model.