SPR1-2674 Fix token tests for PyJWT >= 2.2.0
In order to avoid a test dependency on the cryptography package, I had
a sneaky way to encode tokens with the "none" algorithm, even though
the header requested "ES256" like a real MeerKAT token. I took the
resultant unsecured JWS and then tacked on a dummy signature of the
expected length of 86 characters, since that is all that matters to
katdal.

Sadly this trick stopped working in PyJWT >= 2.2.0, thanks to
[jpadilla/pyjwt#673](https://github.com/jpadilla/pyjwt/pull/673).
The latest PyJWT prefers the algorithm in the header and won't let
it be overridden by the `algorithm` parameter to `jwt.encode()`.

Instead of inventing more tricks, take the hit and add cryptography to
the test requirements. Use the example private key found in the PyJWT
unit tests. There is no need for the corresponding public key, since
katdal only checks some basic aspects of the token without verifying
the signature.