SPR1-2674 Avoid indexing a zero-sized array
Since NumPy 1.20 indexing errors will be reported as DeprecationWarnings
even when the index result is empty. The reason we have an out-of-bound
index is because dask blockwise calls `da.core.compute_meta` on our
block function, which is `_correct_autocorr_quant()` in the Van Vleck
correction `correct_autocorr_quantisation`. This feeds the function
a visibility array of shape (0, 0, 0) to figure out the output array
characteristics. Inside the function we then index the final correlation
product axis to extract the autocorrelations but this now has length 0,
causing an index error.

Fix this by explicitly providing the meta, so that there is no need to
calculate it. Also fix the dtype to be the same as that of vis (it
should be np.complex64 but actually it will preserve vis's dtype).

Pytest reports: 211 passed, 10 skipped, 2 warnings in 59.23s