SPR1-2674 All pytest.raises are context managers
Some of the `pytest.raises` were implemented as functions. Convert all
of them to be context managers in with-statements, as recommended by
the pytest docs.

Add the `match` parameter to `pytest.raises` to clarify the reason
behind each `InvalidToken`. This way I discovered that the expiry time
*can* be a float (since it can be cast to int) but not a floating-point
number as a string. Adjust the test accordingly.