SPR1-2674 Reimplement nose.tools.timed function
This is a useful tool to check that the S3 unit tests don't take longer
than expected. I could not find a pytest alternative (pytest-timeout
is not the timeout I was looking for :-)), so I appropriated the version
in nose, just stripping out the nose-specific `make_decorator` call.

This finally removes nose as a dependency.