SPR1-2674 Rework timed to include minimum duration
Now that the `timed` function is under our control, extend it to check
both the minimum and maximum test duration. The default maximum is
a small tolerance bigger than the minimum, which effectively means
that we are usually checking a specific expected duration (which is
what we want).

Ironically the nose `timed` function originally had the signature
`timed(under, over, exact)` more than 16 years ago, so we are partially
back there again...