SPR1-2674 Don't write 0 bytes to avoid broken pipe
Quit early if there is no response data to write back to the client to
avoid broken pipes. The client typically stops listening if it knows
that nothing more is coming, like in a header-only PUT response.

The broken pipe condition is hard to reproduce... It fails several
times on production Jenkins, but I cannot reproduce it on my laptop
or in an equivalent jenkins-gpu-agent Docker container.