SPR1-2674 Use more robust MinIO health check
The "liveness" (`/minio/health/live`) and "readiness"
(`/minio/health/ready`) checks only confirm that the server is up,
not that it is ready for requests. There is a short period of time
where these checks return 200 (OK) but any actual S3 requests return
503 (Server not initialized). During this time the health response
header contains the following additional key::

    {'X-Minio-Server-Status': 'offline'}

The rationale is that a cluster of MinIO servers might be in the process
of starting up, and the request is then proxied to neighbouring peers
instead of failing fast (see minio/minio#10790).

Check the response header to ensure that the server is initialised
before proceeding.

I also tried the `/minio/health/cluster/read` health endpoint for this
purpose, which worked fine on my laptop but not on Jenkins (probably
due to the older version of MinIO in that environment: 2021-01-08).