SPR1-2674 Undo Jenkins tweaks on duration checks
While Jenkins struggled with the duration checks, I arbitrarily
relaxed the criteria to make the tests pass. Now that Jenkins does
not check test durations anymore, it is safe to revert to the original
values. These pass on my laptop.

I increased the tolerance a bit to avoid the odd failure. Also,
`test_persistent_server_errors` actually fails after 1.0 seconds as
explained in the comments, and not 0.9 seconds as it was before.