SPR1-2674 Redo duration verification with hooks
Be more pytest-onic and use a hook function instead of a standard
decorator with a yucky global variable. The TestReport object also has
a `duration` attribute, so there is no need to do the timing ourselves.

The decorator is replaced by a pytest marker that stores the expected
duration for those tests that need it. Since we will only use
--check-durations in a controlled environment on an ad hoc basis, it is
safe to reduce the tolerance to 0.1 second and also to standardise the
mark info as a single (minimum) duration.

For the `test_store_unavailable_unresponsive_server` test it now means
that we pick the expected duration based on the platform, instead of
allowing a wider time range, which is more accurate.