SPR1-2460 Add basic --scans and --targets options
The --scans option is currently meant for simple 'track' and 'slew'
cases.

The --targets option ran into the issue that katdal messes up the case
where two targets have the same name, only selecting one of them...
Add a hack to fix this for now, but this should go into DataSet proper.